Added test case and refactoring for the folder
