+++ app/utils/non_code_parsing/file_size_checker.py
